create new VehiclesAdjustment class, revert to old UniformVehiclesAdjustment
