+++ src/main/scala/beam/sim/vehicles/UniformVehiclesAdjustment.scala
@@ -9,13 +9,19 @@ import org.matsim.api.core.v01.Coord
 
 case class UniformVehiclesAdjustment(beamScenario: BeamScenario) extends VehiclesAdjustment {
 
-  private val vehicleTypesAndProbabilityByCategoryAndGroup =
-    scala.collection.mutable.Map[(VehicleCategory, String), Array[(BeamVehicleType, Double)]]()
-  beamScenario.vehicleTypes.values.groupBy(x => x.vehicleCategory).map {
-    case (cat, vehTypes) =>
-      val submap = getCategoryAndGroup(cat, vehTypes.toArray)
-      vehicleTypesAndProbabilityByCategoryAndGroup ++= submap
-  }
+  private val vehicleTypesAndProbabilitiesByCategory: Map[(VehicleCategory, String), Array[(BeamVehicleType, Double)]] =
+    beamScenario.vehicleTypes.values.groupBy(x => (x.vehicleCategory, matchCarUse(x.id.toString))).map {
+      case (cat, vehTypes) =>
+        val probSum = vehTypes.map(_.sampleProbabilityWithinCategory).sum
+        val cumulativeProbabilities = vehTypes
+          .map(_.sampleProbabilityWithinCategory / probSum)
+          .scan(0.0)(_ + _)
+          .drop(1)
+          .toList :+ 1.0
+        val vehTypeWithProbability =
+          vehTypes.zip(cumulativeProbabilities).map { case (vehType, prob) => (vehType, prob) }.toArray
+        (cat, vehTypeWithProbability)
+    }
 
   override def sampleVehicleTypesForHousehold(
     numVehicles: Int,
@@ -24,18 +30,9 @@ case class UniformVehiclesAdjustment(beamScenario: BeamScenario) extends Vehicle
     householdSize: Int,
     householdPopulation: Population,
     householdLocation: Coord,
-    realDistribution: UniformRealDistribution,
+    realDistribution: UniformRealDistribution
   ): List[BeamVehicleType] = {
-    val categoryAndGroup = getHouseholdIncomeGroup(householdIncome)
-    val vehTypeWithProbabilityOption = vehicleTypesAndProbabilityByCategoryAndGroup.get(categoryAndGroup)
-    val vehTypeWithProbability: Array[(BeamVehicleType, Double)] = vehTypeWithProbabilityOption match {
-      case Some(vtWithProb) => vtWithProb
-      case _ =>
-        logger.warn(
-          s"There is no vehicle defined for group ${categoryAndGroup._2}, defaulting to ${vehicleTypesAndProbabilityByCategoryAndGroup.head._1._2}"
-        )
-        vehicleTypesAndProbabilityByCategoryAndGroup.head._2
-    }
+    val vehTypeWithProbability = vehicleTypesAndProbabilitiesByCategory(vehicleCategory, "Usage Not Set")
     (1 to numVehicles).map { _ =>
       val newRand = realDistribution.sample()
       val (vehType, _) = vehTypeWithProbability.find { case (_, prob) => prob >= newRand }.get
@@ -48,16 +45,10 @@ case class UniformVehiclesAdjustment(beamScenario: BeamScenario) extends Vehicle
     vehicleCategory: VehicleCategory,
     realDistribution: UniformRealDistribution
   ): List[BeamVehicleType] = {
-    val vehTypeWithProbabilityOption =
-      vehicleTypesAndProbabilityByCategoryAndGroup.get((beam.agentsim.agents.vehicles.VehicleCategory.Car, "ridehail"))
-    val vehTypeWithProbability: Array[(BeamVehicleType, Double)] = vehTypeWithProbabilityOption match {
-      case Some(vtWithProb) => vtWithProb
-      case _ =>
-        logger.warn(
-          s"There is no vehicle defined for ridehail vehicles, defaulting to ${vehicleTypesAndProbabilityByCategoryAndGroup.head._1._2}"
-        )
-        vehicleTypesAndProbabilityByCategoryAndGroup.head._2
-    }
+    val vehTypeWithProbability = vehicleTypesAndProbabilitiesByCategory.getOrElse(
+      (vehicleCategory, "Ride Hail Vehicle"),
+      vehicleTypesAndProbabilitiesByCategory(vehicleCategory, "Usage Not Set")
+    )
     (1 to numVehicles).map { _ =>
       val newRand = realDistribution.sample()
       val (vehType, _) = vehTypeWithProbability.find { case (_, prob) => prob >= newRand }.get
@@ -65,46 +56,11 @@ case class UniformVehiclesAdjustment(beamScenario: BeamScenario) extends Vehicle
     }.toList
   }
 
-  private def getHouseholdIncomeGroup(
-    householdIncome: Double
-  ): (VehicleCategory, String) = {
-    val incomeBin = householdIncome match {
-      case inc if inc <= 17000                 => "inc17"
-      case inc if inc > 17000 && inc <= 36000  => "inc36"
-      case inc if inc > 36000 && inc <= 61000  => "inc61"
-      case inc if inc > 61000 && inc <= 86000  => "inc86"
-      case inc if inc > 86000 && inc <= 133000 => "inc133"
-      case inc if inc > 133000                 => "inc327"
+  private def matchCarUse(vehicleTypeId: String): String = {
+    vehicleTypeId.toString.split("_").headOption match {
+      case Some(beamScenario.beamConfig.beam.agentsim.agents.rideHail.initialization.procedural.vehicleTypePrefix) =>
+        "Ride Hail Vehicle"
+      case _ => "Usage Not Set"
     }
-    (beam.agentsim.agents.vehicles.VehicleCategory.Car, incomeBin)
   }
-
-  def getCategoryAndGroup(
-    category: VehicleCategory,
-    vehTypes: Array[BeamVehicleType]
-  ): scala.collection.mutable.Map[(VehicleCategory, String), Array[(BeamVehicleType, Double)]] = {
-    var groupIDs = scala.collection.mutable.Map[(VehicleCategory, String), Array[(BeamVehicleType, Double)]]()
-    var groupIDlist: scala.collection.mutable.ListBuffer[(VehicleCategory, String)] =
-      scala.collection.mutable.ListBuffer()
-    var vehicleTypeAndProbabilityList: scala.collection.mutable.ListBuffer[(BeamVehicleType, Double)] =
-      scala.collection.mutable.ListBuffer()
-    vehTypes.foreach { vehType =>
-      vehType.sampleProbabilityString.getOrElse("All").replaceAll("\\s", "").toLowerCase.split(";").foreach { group =>
-        val groupAndProbability = group.split(":")
-        if (groupAndProbability.length == 2) {
-          groupIDlist += ((category, groupAndProbability(0)))
-          vehicleTypeAndProbabilityList += ((vehType, groupAndProbability(1).toDouble))
-        }
-      }
-      groupIDlist.zip(vehicleTypeAndProbabilityList).groupBy(_._1).map {
-        case (groupID, vehicleTypeAndProbability) =>
-          val probSum = vehicleTypeAndProbability.map(_._2._2).sum
-          val cumulativeProbabilities = vehicleTypeAndProbability.map(_._2._2 / probSum).scan(0.0)(_ + _).drop(1)
-          val vehicleTypes = vehicleTypeAndProbability.map(_._2._1)
-          groupIDs += (groupID -> vehicleTypes.zip(cumulativeProbabilities).toArray)
-      }
-    }
-    groupIDs
-  }
-
 }
